Time: 7 ms (41.22%), Space: 55.7 MB (16.03%) - LeetHub

+++ 3068-find-the-maximum-sum-of-node-values/3068-find-the-maximum-sum-of-node-values.java
@@ -3,15 +3,17 @@ class Solution {
         int n = nums.length;
         int[] diff = new int[n];
         long ans = 0;
+        
         for(int i=0; i<n; ++i){
             diff[i] = (nums[i]^k) - nums[i];
-            ans += (long)nums[i];
+            ans += nums[i];
         }
+        
         Arrays.sort(diff);
-        // System.out.println(Arrays.toString(diff));
+        
         for(int i=n-1; i>0; i-=2){
             if(diff[i] + diff[i-1] > 0)
-                ans += (long)diff[i] + (long)diff[i-1];
+                ans += diff[i] + diff[i-1];
             else
                 break;
         }
